+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Change Log
 
-fixed undo/redo stack being duplicated at each adition, wich caused undo and redo not work properly
+- added support for tall patches
+- added suppport for color conversion modes when converting png to lmp
+- faster png to lmp using cache for processed colors
+- multi-file conversion added